reorganized file structure and file types
