+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.scss';
 import {useState, useEffect} from 'react'
-import Competition from './components/Competition';
-import Leaderboard from './components/Leaderboard';
+import Competition from './pages/competition/index';
+import Leaderboard from './pages/leaderboards/index';
 import data from './testData/competitorData.json'
 import { Route } from 'react-router-dom'
 import { Routes } from 'react-router';
@@ -9,7 +9,7 @@ import Header from './components/Header';
 import './styles/base.scss'
 import './styles/main.scss'
 import './styles/footer.scss'
-import Competitors from './components/Competitors';
+import Competitors from './pages/competitors/index';
 import apis from './api'
 
 // const schedule = require('node-schedule');
