+++ src/components/modals/VoteConfirmationModal.jsx
