+++ src/pages/competition/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Competition from './Competition';
+import Metrics from './Metrics';
+
+function index() {
+    return (
+        <div>
+            <Metrics />
+            <Competition />
+        </div>
+    );
+}
+
+export default index;
