+++ src/pages/competitors/Competitors.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import api from '../api'
+import api from '../../../api'
 
 function Competitors() {
     const [competitorsState, setCompetitorsState] = useState([])
