+++ src/pages/competitors/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function index() {
+    return (
+        <div>
+            
+        </div>
+    );
+}
+
+export default index;
